Renamed some classes
In the math operations state, was getting the OP2 int value instead of float value, caused all operations of OP1 int divided by OP2 frac < 1 to cause error cause it rounded down to 0. Fixed by calling OP2 float value instead

+++ src/calc/MathOperationsState.java
@@ -30,7 +30,7 @@ public class MathOperationsState
                 total = multiply(CalcSM.OP1.getBigDecimal(), CalcSM.OP2.getBigDecimal());
                 break;
             case DIVIDE:
-                if (CalcSM.OP2.getBigDecimal().intValue() == 0)
+                if (CalcSM.OP2.getBigDecimal().floatValue() == 0)
                 {
                     CalcSM.hasError = true;
                     break;
